fretes

+++ AutoSavePI4/Entidades/FRETES.cs
@@ -9,18 +9,16 @@ namespace AutoSavePI4.Entidades
     public class Fretes
     {
         [Key]
-        public long Codigo { get; set; }
+        public int Codigo { get; set; }
         public string? TipoCarga { get; set; }
         public string? Pesocarga { get; set; }
         public string? LocalColeta { get; set; }
         public string? LocalEntrega { get; set; }
         public decimal? ValorPagamento { get; set; }
-        public long? SeqEmpresa { get; set; }
-        public long? SeqUsuario { get; set; }
+        public int? SeqEmpresa { get; set; }
+        public int? SeqUsuario { get; set; }
         public Empresa? Empresa { get; set; }
         public Usuario? Usuario { get; set; }
-
-
 
     }
 }
